+++ src/Twig/RadTwigExtension.php
@@ -4,6 +4,7 @@ namespace Torr\Rad\Twig;
 
 use Torr\Rad\Html\DataContainer;
 use Twig\Extension\AbstractExtension;
+use Twig\TwigFilter;
 use Twig\TwigFunction;
 
 class RadTwigExtension extends AbstractExtension
@@ -20,12 +21,34 @@ class RadTwigExtension extends AbstractExtension
 
 
 	/**
+	 * Appends the given string value to the $array $key (and creates the key if it didn't exist)
+	 */
+	public function appendToArrayKey (array $array, string $key, string $value) : array
+	{
+		$current = $array[$key] ?? "";
+		$array[$key] = \trim("{$current} {$value}");
+		return $array;
+	}
+
+
+	/**
 	 * @inheritDoc
 	 */
 	public function getFunctions () : array
 	{
 		return [
 			new TwigFunction("data_container", [$this->dataContainer, "renderToHtml"], ["is_safe" => ["html"]]),
+		];
+	}
+
+
+	/**
+	 * @inheritDoc
+	 */
+	public function getFilters () : array
+	{
+		return [
+			new TwigFilter("appendToArrayKey", [$this, "appendToArrayKey"]),
 		];
 	}
 }