+++ CHANGELOG.md
@@ -7,3 +7,4 @@
     *   `IdTrait`, `SortOrderTrait`, `TimestampsTrait`
 *   (feature) Add `DataContainer`.
 *   (feature) Add Twig function: `data_container`
+*   (feature) Add Twig filter: `appendToArrayKey`